Fix EDNS0 UDP Payload Size Clamping
Forwarded queries take the EDNS0 UDP Payload Size provided by the client. This can result in forwarded queries that exceed the save ENDS0 UDP Payload Size as well as failed replies due to UDP fragmentation reconstruction errors or dropped packets.

This clamps the EDNS0 Payload Size on forwarded queries to be the lower of the client provided value or --edns-packet-max (default 1232) to ensure reliable delivery.